Move configure/ into parse/; rename ReadFiles to Read
parse now owns the full raw-data → pipeline-descriptor pipeline:
- parse.Read(path) replaces configure.ReadFiles (same logic, cleaner name)
- parse/hcl/ replaces configure/hcl/ (import path update only)
- configure/ package removed
